loss

+++ model/loss.py
@@ -333,8 +333,8 @@ class CaptGANDiscriminatorLoss(torch.nn.Module):
 
     def forward(self, evaluator_outputs, generator_outputs, other_outputs):
         batch_size = evaluator_outputs.size(0)
-        real_labels = torch.ones(batch_size, 1).type(torch.LongTensor)
-        fake_labels = torch.zeros(batch_size, 1).type(torch.LongTensor)
+        real_labels = torch.ones(batch_size, 1)
+        fake_labels = torch.zeros(batch_size, 1)
         real_labels = Variable(real_labels).to(self.device)
         fake_labels = Variable(fake_labels).to(self.device)
